feat(safestep): fixed map rendering
fixed map rendering
