+++ packages/mobile/src/app/(private)/recoverer/map.tsx
@@ -12,22 +12,6 @@ const CurrentLocationMap = () => {
 
   console.log("current location map geofencing state:", geofencingState);
 
-  const HARDCODED_LATITUDE = 41.48595203395239;
-  const HARDCODED_LONGITUDE = -71.42164970467098;
-
-  // Mock the LocationObjectCoords type expected by LocationMap
-  const mockCurrentLocationCoords = {
-    latitude: HARDCODED_LATITUDE,
-    longitude: HARDCODED_LONGITUDE,
-    // Add other properties that the LocationObjectCoords type requires,
-    // even if they are zero or null, to satisfy the type definition.
-    accuracy: 5,
-    altitude: 0,
-    altitudeAccuracy: 0,
-    heading: 0,
-    speed: 0,
-  };
-
   useEffect(() => {
     startTracking();
     startGeofencing();
@@ -37,12 +21,7 @@ const CurrentLocationMap = () => {
 
   return (
     trackingState.state === "tracking" &&
-    isFocused && (
-      <LocationMap
-        currentLocationCoords={mockCurrentLocationCoords}
-        geofencingState={geofencingState}
-      />
-    )
+    isFocused && <LocationMap geofencingState={geofencingState} />
   );
 };
