--- packages/mobile/src/app/(private)/recoverer/recovererMap.tsx
@@ -1,7 +1,0 @@
-import LocationMap from "@/components/custom/LocationMap";
-
-const recovererMap = () => {
-  return <LocationMap></LocationMap>;
-};
-
-export default recovererMap;
