+++ packages/mobile/src/app/(private)/sponsor/_layout.tsx
@@ -1,5 +1,4 @@
 import { Tabs } from "expo-router";
-import { StatusBar } from "expo-status-bar";
 import FontAwesome from "@expo/vector-icons/FontAwesome";
 
 const SponsorLayout = () => {
